point maze env
