+++ gymnasium_robotics/envs/point_maze/maze_env.py
@@ -1,0 +1,275 @@
+import tempfile
+import xml.etree.ElementTree as ET
+from os import path
+from typing import Dict, Optional
+
+import numpy as np
+
+from gymnasium_robotics.core import GoalEnv
+from gymnasium_robotics.envs.ant_maze.maps import COMBINED, GOAL, RESET, U_MAZE
+
+
+class Maze:
+    def __init__(self, maze_map: list, maze_size_scaling: float, maze_height: float):
+
+        self._maze_map = maze_map
+        self._maze_size_scaling = maze_size_scaling
+        self._maze_height = maze_height
+
+        self._unique_goal_locations = []
+        self._unique_reset_locations = []
+        self._combined_locations = []
+
+        # Get the center cell position of the maze. This will be the origin
+        self._map_length = len(maze_map)
+        self._map_width = len(maze_map[0])
+        self._x_map_center = np.ceil(self.map_width / 2) * maze_size_scaling
+        self._y_map_center = np.ceil(self.map_length / 2) * maze_size_scaling
+
+    @property
+    def maze_map(self):
+        return self._maze_map
+
+    @property
+    def maze_size_scaling(self):
+        return self._maze_size_scaling
+
+    @property
+    def maze_height(self):
+        return self._maze_height
+
+    @property
+    def unique_goal_locations(self):
+        return self._unique_goal_locations
+
+    @property
+    def unique_reset_locations(self):
+        return self._unique_reset_locations
+
+    @property
+    def combined_locations(self):
+        return self._combined_locations
+
+    @property
+    def map_length(self):
+        return self._map_length
+
+    @property
+    def map_width(self):
+        return self._map_width
+
+    @property
+    def x_map_center(self):
+        return self._x_map_center
+
+    @property
+    def y_map_center(self):
+        return self._y_map_center
+
+    def cell_rowcol_to_xy(self, rowcol_pos: np.ndarray) -> np.ndarray:
+        x = rowcol_pos[1] - self.x_map_center
+        y = rowcol_pos[0] - self.y_map_center
+
+        return np.array([x, y])
+
+    @classmethod
+    def make_maze(
+        cls,
+        agent_xml_path: str,
+        maze_map: list,
+        maze_size_scaling: float,
+        maze_height: float,
+    ):
+        tree = ET.parse(agent_xml_path)
+        worldbody = tree.find(".//worldbody")
+
+        maze = cls(maze_map, maze_size_scaling, maze_height)
+
+        for i in range(maze.map_length):
+            for j in range(maze.map_width):
+                struct = maze_map[i][j]
+                # Store cell locations in simulation global Cartesian coordinates
+                x = j * maze_size_scaling - maze.x_map_center
+                y = i * maze_size_scaling - maze.y_map_center
+                if struct == 1:  # Unmovable block.
+
+                    # Offset all coordinates so that maze is centered.
+                    ET.SubElement(
+                        worldbody,
+                        "geom",
+                        name=f"block_{i}_{j}",
+                        pos=f"{x} {y} {maze_height / 2 * maze_size_scaling}",
+                        size=f"{0.5 * maze_size_scaling} {0.5 * maze_size_scaling} {maze_height / 2 * maze_size_scaling}",
+                        type="box",
+                        material="",
+                        contype="1",
+                        conaffinity="1",
+                        rgba="0.7 0.5 0.3 1.0",
+                    )
+
+                elif maze_map[i][j] in [
+                    RESET,
+                ]:
+                    print("RESET")
+                    maze._unique_reset_locations.append(np.array([x, y]))
+                elif maze_map[i][j] in [
+                    GOAL,
+                ]:
+                    maze._unique_goal_locations.append(np.array([x, y]))
+                elif maze_map[i][j] in [
+                    COMBINED,
+                ]:
+                    maze._combined_locations.append(np.array([x, y]))
+
+        # Add target site for visualization
+        ET.SubElement(
+            worldbody,
+            "site",
+            name="target",
+            pos=f"0 0 {maze_height / 2 * maze_size_scaling}",
+            size=f"{0.2 * maze_size_scaling}",
+            rgba="1 0 0 0.7",
+            type="sphere",
+        )
+
+        # Add the combined cell locations (goal/reset) to goal and reset
+        maze._unique_goal_locations += maze._combined_locations
+        maze._unique_reset_locations += maze._combined_locations
+
+        # Save new xml with maze to a temporary file
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            temp_xml_path = path.join(path.dirname(tmp_dir), "ant_maze.xml")
+            tree.write(temp_xml_path)
+
+        return maze, temp_xml_path
+
+
+class MazeEnv(GoalEnv):
+    def __init__(
+        self,
+        agent_xml_path,
+        reward_type="dense",
+        continuing_task=True,
+        maze_map=U_MAZE,
+        maze_size_scaling=1,
+        maze_height=0.5,
+        position_noise_range: float = 0.25,
+        **kwargs,
+    ):
+
+        self.reward_type = reward_type
+        self.continuing_task = continuing_task
+        self.maze, self.tmp_xml_file_path = Maze.make_maze(
+            agent_xml_path, maze_map, maze_size_scaling, maze_height
+        )
+
+        self.postion_noise_range = position_noise_range
+
+    def generate_target_goal(self):
+        assert len(self.maze.unique_goal_locations) > 0
+        goal_index = self.np_random.integers(
+            low=0, high=len(self.maze.unique_goal_locations)
+        )
+        goal = self.maze.unique_goal_locations[goal_index].copy()
+
+        return goal
+
+    def generate_reset_pos(self):
+        assert len(self.maze.unique_reset_locations) > 0, ""
+
+        # While reset position is close to goal position
+        reset_pos = self.goal.copy()
+        while np.linalg.norm(reset_pos - self.goal) <= 0.5:
+            reset_index = self.np_random.integers(
+                low=0, high=len(self.maze.unique_reset_locations)
+            )
+            reset_pos = self.maze.unique_reset_locations[reset_index].copy()
+
+        return reset_pos
+
+    def reset(
+        self,
+        *,
+        seed: Optional[int] = None,
+        options: Dict[str, Optional[np.ndarray]] = {
+            "goal_cell": None,
+            "reset_cell": None,
+        },
+    ):
+        super().reset(seed=seed)
+
+        if options["goal_cell"] is not None:
+            # assert that goal cell is valid
+            assert self.maze.map_length > options["goal_cell"][1]
+            assert self.maze.map_width > options["goal_cell"][0]
+            assert (
+                self.maze.maze_map[options["goal_cell"][1], options["goal_cell"][0]]
+                != 1
+            ), f"Goal can't be placed in a wall cell, {options['goal_cell']}"
+
+            goal = self.maze.cell_rowcol_to_xy(options["goal_cell"])
+
+        else:
+            goal = self.generate_target_goal()
+
+        # Add noise to goal position
+        self.goal = self.add_xy_position_noise(goal)
+
+        # Update the position of the target site for visualization
+        self.update_target_site_pos()
+
+        if options["reset_cell"] is not None:
+            # assert that goal cell is valid
+            assert self.maze.map_length > options["reset_cell"][1]
+            assert self.maze.map_width > options["reset_cell"][0]
+            assert (
+                self.maze.maze_map[options["reset_cell"][1], options["reset_cell"][0]]
+                != 1
+            ), f"Reset can't be placed in a wall cell, {options['reset_cell']}"
+
+            reset_pos = self.maze.cell_rowcol_to_xy(options["reset_cell"])
+
+        else:
+            reset_pos = self.generate_reset_pos()
+
+        # Add noise to reset position
+        self.reset_pos = self.add_xy_position_noise(reset_pos)
+
+    def add_xy_position_noise(self, xy_pos: np.ndarray) -> np.ndarray:
+        noise_x = (
+            self.np_random.uniform(
+                low=-self.postion_noise_range, high=self.postion_noise_range
+            )
+            * self.maze.maze_size_scaling
+        )
+        noise_y = (
+            self.np_random.uniform(
+                low=-self.postion_noise_range, high=self.postion_noise_range
+            )
+            * self.maze.maze_size_scaling
+        )
+        xy_pos[0] += noise_x
+        xy_pos[1] += noise_y
+
+        return xy_pos
+
+    def compute_reward(self, achieved_goal, desired_goal, info):
+        if self.reward_type == "dense":
+            return np.exp(-np.linalg.norm(desired_goal - achieved_goal))
+        elif self.reward_type == "sparse":
+            return 1.0 if np.linalg.norm(achieved_goal - desired_goal) <= 0.5 else 0.0
+
+    def compute_terminated(self, achieved_goal, desired_goal, info):
+        if not self.continuing_task:
+            # If task is episodic terminate the episode when the goal is reached
+            return bool(np.linalg.norm(achieved_goal - desired_goal) <= 0.5)
+        else:
+            # Continuing tasks don't terminate, episode will be truncated when time limit is reached (`max_episode_steps`)
+            return False
+
+    def compute_truncated(self, achieved_goal, desired_goal, info):
+        return False
+
+    def update_target_site_pos(self, pos):
+        """If the"""
+        raise NotImplementedError
