+++ gymnasium_robotics/envs/point_maze/point_env.py
@@ -1,0 +1,66 @@
+from os import path
+from typing import Optional
+
+import numpy as np
+from gymnasium import spaces
+from gymnasium.envs.mujoco.mujoco_env import MujocoEnv
+
+
+class PointEnv(MujocoEnv):
+
+    metadata = {
+        "render_modes": [
+            "human",
+            "rgb_array",
+            "depth_array",
+        ],
+        "render_fps": 100,
+    }
+
+    def __init__(self, xml_file: Optional[str] = None, **kwargs):
+
+        if xml_file is None:
+            xml_file = path.join(
+                path.dirname(path.realpath(__file__)), "../assets/point/point.xml"
+            )
+        observation_space = spaces.Box(
+            low=-np.inf, high=np.inf, shape=(4,), dtype=np.float64
+        )
+        super().__init__(
+            model_path=xml_file,
+            frame_skip=1,
+            observation_space=observation_space,
+            **kwargs
+        )
+
+    def reset_model(self):
+        self.set_state(self.init_qpos, self.init_qvel)
+        obs = self._get_obs()
+
+        return obs
+
+    def step(self, action):
+        action = np.clip(action, -1.0, 1.0)
+        self._clip_velocity()
+        self.do_simulation(action, self.frame_skip)
+        obs, info = self._get_obs()
+        # This environment class has no intrinsic task, thus episodes don't end and there is no reward
+        reward = 0
+        terminated = False
+        truncated = False
+
+        if self.render_mode == "human":
+            self.render()
+
+        return obs, reward, terminated, truncated, info
+
+    def _get_obs(self):
+        return np.concatenate([self.data.qpos, self.data.qvel]).ravel(), {}
+
+    def _clip_velocity(self):
+        qvel = np.clip(self.data.qvel, -5.0, 5.0)
+        self.set_state(self.data.qpos, qvel)
+
+    def viewer_setup(self):
+        assert self.viewer is not None
+        self.viewer.cam.distance = self.model.stat.extent
